requirements.txt
discord.py
python-dotenv
imapclient
beautifulsoup4
flask
gunicorn
